requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.3
validators==0.33.0
